[TNT] Refactor tank accessory types
Tank attachments and tap requirements, collectively referred to as tank
accessories, are very similar both in concept and infrastructure. So
similar that I've been referring to "tank accessories" for some time
now, even though that concept did not appear anywhere in the code: the
code contained only separate attachments and requirements.

In this commit, these two types are somewhat unified into a single list
of tank accessories, which are then subclassified into attachments or
requirements. This simplifies infrastructure significantly, as an
accretion of nearly duplicate code was forming to handle first the
attachments, then the requirements. Now, code which wishes to treat the
two as a single class (tank accessories) can do so, while code which
wishes to handle them separately can do so as well.