[TNT] Remove asset_store class again
It was fun while it lasted, but the results are in and this type is
simply too pedantic to work in our codebase. It makes no sense to do
such careful checking of balance manipulations if you allow live
balances to be copied rather than moved; however, copying is critical
for our undo infrastructure, so unfortunately, the asset_store class is
a nonstarter.

This reverts commit aa455163b5ed620e8181b5a2b9f4764a5ff83a15.