[TNT] Replace deposit_source_restrictor
The deposit_source_restrictor has been removed in favor of the
ubiquitous authorized_connections_type. This type is simpler and easier
to reason about, as it does not involve pattern matching but simply
lists out the sources which are allowed to connect to a connection
target. This also improves security by making deposit source checking a
universal concern rather than relegating it to the
deposit_source_restrictor.